+++ src/api/API.js
@@ -1,0 +1,2 @@
+export const ROCKETS_URL = 'https://api.spacexdata.com/v3/rockets';
+export const MISSIONS_URL = 'https://api.spacexdata.com/v3/missions';
