+++ src/routes/Navbar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import logo from '../../assets/img/logo.png';
+import logo from '../assets/img/logo.png';
 import styles from './Navbar.module.css';
 
 const Navbar = () => (
